@@ -14,20 +14,21 @@
     "storybook": "start-storybook"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.16",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.0",
-    "@angular/cli": "^7.3.0",
-    "@angular/common": "^7.2.16",
-    "@angular/compiler": "^7.2.16",
-    "@angular/compiler-cli": "^7.2.16",
-    "@angular/core": "^7.2.16",
-    "@angular/forms": "^7.2.16",
-    "@angular/http": "^7.2.16",
-    "@angular/platform-browser": "^7.2.16",
-    "@angular/platform-browser-dynamic": "^7.2.16",
-    "@angular/router": "^7.2.16",
+    "@angular/cli": "^8.3.25",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@cyber4all/clark-taxonomy": "^4.0.1",
+    "autoprefixer": "^8.6.5",
     "aws-sdk": "^2.607.0",
+    "babel-core": "^6.26.3",
     "chance": "^1.1.4",
     "chart.js": "^2.9.3",
     "core-js": "^2.6.11",
@@ -45,11 +46,11 @@
     "rxjs-compat": "^6.5.4",
     "tslib": "^1.9.0",
     "uuid": "^3.4.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/language-service": "^7.2.16",
+    "@angular-devkit/build-angular": "~0.803.25",
+    "@angular/language-service": "^8.2.14",
     "@babel/core": "^7.8.3",
     "@storybook/angular": "^5.3.8",
     "@types/aws-sdk": "^2.7.0",
@@ -60,16 +61,16 @@
     "babel-loader": "^8.0.6",
     "canvas": "^2.6.1",
     "chrome-launcher": "^0.10.7",
-    "codelyzer": "^4.3.0",
+    "codelyzer": "^5.0.1",
     "install-peers": "^1.0.2",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^24.9.0",
+    "jest": "^23.6.0",
     "jest-preset-angular": "^6.0.2",
     "lighthouse": "^5.6.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.10.0",
-    "typescript": "3.2.4"
+    "typescript": "3.5.3"
   },
   "prettier": {
     "singleQuote": true
